use unique_ptr for labels
