+++ core/src/labels/labelMesh.h
@@ -1,42 +1,27 @@
 #pragma once
 
 #include "gl/typedMesh.h"
-//#include "labels/label.h"
-#include "glm/vec2.hpp"
+#include "labels/label.h"
 
 #include <memory>
 #include <vector>
 
 namespace Tangram {
 
-class Label;
-
-struct BufferVert {
-    glm::vec2 pos;
-    glm::vec2 uv;
-    struct State {
-        glm::vec2 screenPos;
-        float alpha;
-        float rotation;
-    } state;
-};
-
-class LabelMesh : public TypedMesh<BufferVert> {
+class LabelMesh : public TypedMesh<Label::BufferVert> {
 public:
-    LabelMesh(std::shared_ptr<VertexLayout> _vertexLayout, GLenum _drawMode)
-        : TypedMesh<BufferVert>(_vertexLayout, _drawMode, GL_DYNAMIC_DRAW) {
-    }
+    LabelMesh(std::shared_ptr<VertexLayout> _vertexLayout, GLenum _drawMode);
 
-    void addLabel(std::shared_ptr<Label> _label) {
-        m_labels.push_back(std::move(_label));
-    }
+    virtual ~LabelMesh();
 
-    std::vector<std::shared_ptr<Label>>& getLabels() {
+    void addLabel(std::unique_ptr<Label> _label);
+
+    std::vector<std::unique_ptr<Label>>& getLabels() {
         return m_labels;
     }
 
 protected:
-    std::vector<std::shared_ptr<Label>> m_labels;
+    std::vector<std::unique_ptr<Label>> m_labels;
 };
 
 }
